+++ src/statusBarMessage.ts
@@ -1,8 +1,9 @@
 import { diagnosticToInlineMessage, isSeverityEnabled } from 'src/decorations';
 import { $config } from 'src/extension';
-import { AggregatedByLineDiagnostics, CommandIds, ExtensionConfig } from 'src/types';
+import { AggregatedByLineDiagnostics, CommandId, ExtensionConfig } from 'src/types';
 import { replaceLinebreaks } from 'src/utils';
-import { Diagnostic, Position, StatusBarItem, TextEditor, ThemeColor, window } from 'vscode';
+import { Diagnostic, Position, StatusBarAlignment, StatusBarItem, TextEditor, ThemeColor, window } from 'vscode';
+
 /**
  * Handle status bar updates.
  */
@@ -32,20 +33,23 @@ export class StatusBarMessage {
 		private readonly isEnabled: boolean,
 		private readonly colorsEnabled: boolean,
 		private readonly messageType: ExtensionConfig['statusBarMessageType'],
+		priority: ExtensionConfig['statusBarMessagePriority'],
+		alignment: ExtensionConfig['statusBarMessageAlignment'],
 	) {
+		const statusBarAlignment = alignment === 'right' ? StatusBarAlignment.Right : StatusBarAlignment.Left;
 		this.colorsEnabled = colorsEnabled;
 		this.messageType = messageType;
 
-		this.statusBarItem = window.createStatusBarItem('errorLensMessage', -9999);
+		this.statusBarItem = window.createStatusBarItem('errorLensMessage', statusBarAlignment, priority);
 		this.statusBarItem.name = 'Error Lens: Message';
-		this.statusBarItem.command = CommandIds.statusBarCommand;
+		this.statusBarItem.command = CommandId.statusBarCommand;
 		if (this.isEnabled) {
 			this.statusBarItem.show();
 		} else {
 			this.dispose();
 		}
 	}
-	updateText(editor: TextEditor, aggregatedDiagnostics: AggregatedByLineDiagnostics) {
+	updateText(editor: TextEditor, aggregatedDiagnostics: AggregatedByLineDiagnostics): void {
 		if (!this.isEnabled) {
 			return;
 		}
@@ -130,7 +134,7 @@ export class StatusBarMessage {
 	/**
 	 * Clear status bar message.
 	 */
-	clear() {
+	clear(): void {
 		if (!this.isEnabled) {
 			return;
 		}
@@ -140,7 +144,7 @@ export class StatusBarMessage {
 	/**
 	 * Dispose status bar item.
 	 */
-	dispose() {
+	dispose(): void {
 		this.statusBarItem.dispose();
 	}
 }